Zero Individual Traces in Pulse Averaging Graph
A checkbox is introduced to the PA section in BSP. The user can select
to zero traces using subsequent Differentiation and Integration.
Calculation is performed in `ZeroWave()`. Before performing a
calculation, a backup wave is created. This backup wave is valid as long
as the source wave does not change in Modification date or the length of
the wave does not change. A wave note entry `NOTE_KEY_ZEROED` keeps track
whether or not the calculation must be performed.

Additional notes
================

* Note that upon Differentiation, the length of the wave is reduced by
two and the length of the wave is therefore stored as entry to the wave
note.

* All individual traces from `listOfWaves` are zeroed.

* `SetNumberInWaveNote` was preferred over `AddEntryIntoWaveNoteAsList`
due to some incompatibilities of the two functions and the more general
use of `SetNumberInWaveNote` throughout the project.

* If the PA traces are rebuilt from scratch, also initialize the Wave
Note entry to indicate that no zeroing has been performed at this point
and discard any backup waves to avoid overwriting with old data.

Original-CommitID: 2add1bbab1de15de9e8c4f398349e443eb13e3de