ExpConfig_Amplifiers: Don't mixup DAC and headstages
Since forever the experiment config code when activating the headstages
did so by activating the associated DA channel and by
DAP_AdaptAssocHeadstageState the headstage got activated as well.

But actually this should not work as for getting the headstage from a DA
channel we need to know the clamp mode of the DA channel but the clamp
mode is a feature of the headstage.

And this does not work anymore since 471f21f3 (GetChannelClampMode:
Initialize wave to NaN, 2019-03-08) and cb206d3c (GetChannelClampMode:
Extend to also hold headstage information, 2019-03-08).

The correct fix here is to activate the headstage by using the headstage
control.

Original-CommitID: 1dd224d12f5e088364678ec99182596093bfd3e9