DA_Ephys: Make it possible again to activate a headstage from the AD/DA channels
Since cb206d3c (GetChannelClampMode: Extend to also hold headstage
information, 2019-03-08) it is not possible anymore to activate a
headstage by clicking on the AD/DA channels.

The reason being that we don't know the clamp mode for a DA/AD channel
if these are not connected.

This is in theory correct but quite inconvenient for users.

Let's restore the old behaviour for the case that that VC/IC channel
number of the headstage is the same.

Original-CommitID: 4ed58c6280a44993d3f5be7c6e716e97c49c5f17