Preserve Cursor Position on Sweep Changes
All traces are deleted when changing sweeps using the Browser Settings
Control Panel actions (previous, next, etc). This also leads to the
the fact that the cursors associated with the trace are deleted. Cursors
are used in Time Alignment for specifying the feature position. As these
feature positions stay roughly the same throughout a typical user
operation, the corresponding cursors will not be reset.

This feature was already tried to accomplish in XXXX but left
unfunctional in current revision.