Add Time Alignment for Pulse Averaging Graph
Timealignment is done for sweeps from the PA graph with the same region
as the selected reference trace. Feature selection is done for all
sweeps on the same axis. Time Alignment only acst on the PA graph or the
main window (SB/DB). There is no option to act on both simultaneously.

The "multiple graphs" option for faster display in OVS mode was tested
and is considered working.

The current trace properties like region and sweepNumer are saved as
userData properties. To get to the calling BSP window from a PA window,
the host window that triggered the built of the PA window (DB or a SB)
is saved as window userdata element.

Changes in the popupmenu always trigger the cursor to switch to the
selected trace. Some new Cursor functions are introduced to handle the
cursor switching between traces.

The popupmenu was switched to a "window-based" syntax introducing a
"graph#name" pattern (named `graphname`) inspired by the naming of
functions in IMs.