AI_QueryGainsFromMCC: Override safety check for holding potential during automated testing
For automated testing we sometimes have the case that the holding
potential in the MCC is enabled.

And in that case we, correctly, don't change the clamp mode if a holding
potential is enabled.

For automated testing we don't need that safety check and by disabling
it we can enhance the reliability of the tests as this should reduce the
amount of manual fixing up.

Original-CommitID: f47e8fe57a05183cd2d903f809e9451dde81bebe