Check If Selected Time Alignment Trace Is Valid
Search the list of traces from the current graph for a valid selection
for Time Alignment.

Matching is done from the currently selected trace for traces plotted in
the same graph and on the same axis as the selected one. The trace list
is searchd downwards for a valid trace using the order from the List
generated by TraceNameList. As average waves are usually added on top of
other traces this allows to select the next valid match after the
average wave. If no trace can be determined, the first available
trace is used. Usually this is the first trace in the main window.