ChangeWaveLock supports now Wave Reference wave with waves storing data. This is used for the NI data waves that store a wave for each channel.
Adapted PSQ_Ramp for NI. DA wave was not properly zeroed as it was locked.

After a spike has been detected the DA wave with NI was not properly zeroed.
While the zeroing was already implemented the wave was locked at that point,
so it failed silently.

Fix: Unlock the wave to zero it and relock it afterwards

Original-CommitID: 25f393e0b679a719b9f16f57c6d820e8d7f08459