NWB Export: Write the correct stimulus set for TTL channels
Since forever we indexed listOfStimsets using the ttlBit, but that is
wrong as it is the log 2 value of it which is the index.

The tests did not catch that as we assumed that we don't add this entry
for TTL data. But in fact we are adding it, and as it works for NI
hardware this was now caught.