PA_ShowPulses: Use full resolution for diagonal traces
In 12484be6 (PA_ShowPulses: Speedup plot creation and resize behaviour,
2019-04-23) we started using a step size > 1 for plotting the PA traces
to speedup the plotting.

Turns out for the diagonal traces this is problematic as the reduced
precision in plotting hinders looking at the data properly.

This is now changed so that we use full resolution for the diagonal ones
and the reduced for off-diagonal.