DAP_CheckHeadStage: Check logic for I=0 case
Since the automated amplifier propagation in 76d6e822 (Propagate
amplifier settings before DAQ/TP, 2016-05-11) the clamp mode I=0 can not
be used for DAQ.

The reason is that the check in the if clause is totally wrong.

The first part needs to ensure that I=0 is not set and the second part
needs to ensure that DAGainMCC is zero, therefore negating the previous
logic is required.