Testpulse: Add fast option for stopping and restarting
For appliations where we need to stop the TP in order to interact with
the amplifier it is beneficial if we can do so as quickly as possible.

For that the new "fast" option is introduced which skips all
unnecessary things.

These are currently:
- Checking the settings via DAP_CheckSettings
- Writing into the labnotebook
- Changing the GUI state

See the "fast" paths in TP_Setup and TP_TearDown for more info.

You can stop in "fast" mode where it was started with "normal" mode, but
if you want to start in "fast" mode it needs to be stopped in "fast"
mode as well.