AI_UpdateAmpModel: Move TP stopping/restarting out of the loop
When "Send to All headstages" is checked we have multiple headstages we
want to do "Auto pipette offset" on. On we used to then always stop and
start the TP for each of them. This is slow even for TP fast mode.

But we can do better and just stop it before the loop and restart it
afterwards.