fix for tests failing due to undefined compile order of functions
UTF_Utils.ipf contained override functions for MIES functions
used in several test cases that depended to be compiled before
MIES itself was compiled.
This is only guaranteed for override functions if they appear in the experiment
procedure window.

The two override functions have been moved to the procedure window of
RunAllTests.pxp