DAP_AdaptAssocHeadstageState: Use correct headstage for quick check
We need to use the headstage we derived from the channel number (idx)
and not the channel number itself.

This bug resulted in an error on ITC1600 hardware with our tests. For
non-ITC1600 hardware this did not result in a bug as we had for all
channel numbers also a corresponding headstage.

Bug introduced in 1449891d (MIES_PanelITC.ipf: Disband file,
2017-11-14).