fix for Test TASYNC_StopTimeOutForce for changes in IP 8.03+
with a nightly build of Igor prior 8.03 the behavior of ThreadGroupRelease
was changed, such that it does not return error -2 which means Igor had to
force the threads to quit.

The fix is that it now passes, even with a timeout of 0.
For Igor version prior 8.03 the old test behavior is kept