initial-repo-config.sh: Make it more robust and work out of the box
For our CI servers we want to have single script which creates a usable
git repository.

For that we need to not only set the clean smudge filters but also
re-checkout the files to be decompressed.

In addition our compression scheme was not 100% reproducible across
platforms. The initial compression was done on a debian stretch with
gzip 1.6 and that seems to differ from the output it creates compared to
gzip 1.10 from git for windows 2.22.0.rc1.windows.1.

As we want a 100% reproducible compression scheme, as otherwise
rebasing is not possible due to "unstaged changes" issues, we set all
parameters of gzip to fixed values. And also use a fixed compression
default of zip level 3.