Require specific IP8 nightly build
We currently have a couple of places where we take advantage of bugfixes
in post 8.03 nightly build versions.

This starts to get unpleasant to rid and debug issues with.

Enfore a tested IP8 nightly build. We don't point to the stock
wavemetrics page as that changes everyday and we want to always have the
same nightly build.