MIES Installer patch for NSIS ShellLink Plug-In
The ShellLink plug-in is used to retrieve the target of a link, like the link
to MIES_Include.ipf in Igor Procedures. Originally it failed if it
could not write to the targeted lnk file.

The patch requests only write access if a operation is used that requires write
, otherwise only read access is requested. This allows to retrieve also link
targets from lnk files written by an admin.

This commit adds the patched shelllink libraries for nsis as well as the
Shelllink plugin project upconverted for use with VS2017.
I had to remove the /SAFESEH linker switch in the project.

Original version from https://nsis.sourceforge.io/ShellLink_plug-in