MIES ASYNC Framework, made ASYNC_Start and ASYNC_Stop idempotent
Both function check now if the framework if already running or not.
If the framework is already in the requested state the functions return.

In such case ASYNC_Start does not change the number of threads or
ReadOut task state.

IsASYNCRunning has some code that prevents the Debugger to popup each time
on compile if a developer has DebugOnError enabled.