Replacement of oodDAQ offset calculation routines
oodDAQ offset calculation is now based on region waves.
Region waves are 2D waves with two column that store begin and end of signal
ranges.

On creation of these regions waves the pre and post delay is taken into account.

The offsets are calculated by finding overlaps of regions and shifting the
current regions by the maximum of overlap found. This is repeated until there is
no overlap.

For subsequent stimsets the offsetted regions get combined. In the case of yoking
the combined regions get stored for the next follower device.

- yoking support included
- removed unused data structures and old functions