WB_MakeWaveBuilderWave: Clarify pulse starting time calculation
Since 0e52306f (WB_MakeWaveBuilderWave: Factor out epochID filling into
WB_UpdateEpochID, 2017-10-17) the pulse starting times for all but the
first epoch are relative to the start of the epoch.

This is how it was designed, according to the commit message in d456a2d5
(Pulse Train: Add the pulse starting times to the wave note,
2017-02-02), but the original code until 0e52306f used stimset build ms
instead.

After internal discussion we concluded that "relative to the epoch
start" is the right thing to do. So let's change the code that this is
obvious, this also removes the difficult to understand
accumulatedDuration updating in WB_UpdateEpochID via a pass-by-ref
parameter.