Wavebuilder: Force stimset rebuilding on too old wave note version
If we ever need to fix bugs in the stimset wave note generation code we
currently have no way of forcing a rebuilding, and thus regeneration of
the stimset wave note.

With the new stimset wave note version check, we can force it by
increasing the wave note version. This also blends in our otherwise used
upgrade strategy for waves.