fix for cases where DA_Ephys panel is not in sync with GUIState waves
There still existed a few situations where it happens that existing
GUIState waves are not in sync with the available controls of the
DA_Ephys panel. e.g. if a dev has changed controls without increasing
the PANEL VERSION constant. Or if an experiment was loaded with an open
panel with a different control state.

The fix is to reset the GUIState waves if
- The panel is locked as this changes the panel title and thus the
  the path to the GUIState waves
- The panel is unlocked as this changes the panel title and thus the
  the path to the GUIState waves. It is reset for the specific Device
  that is unlocked as well as for "DA_Ephys".