Code cleanup: Replace DAorTTL with channelType
Since forever the variables DAOrTTL were used to signal that we have
either DA or TTL channels. As it was never obvious that DAorTTL == 0
means DA and DAorTTL == 1 be TTL, the constants CHANNEL_TYPE_DAC and
CHANNEL_TYPE_TTL where created in 4e749e75 (Calculate maxITI and respect
indexing (locked and unlocked), 2014-10-09) with these values.

And this commit now removes the last occurences of DAorTTL and replaces
it with channelType.