DAP_SetVarProc_Channel_Search: Fix search string application
Since forever the search string worked on the stimset waves directly.
This had the obvious drawback that it requires the stimsets to be
already created when we filter them.

This is of course usually not the case. In addition the current approach
differs from the way we generate the list of stimsets, as
ReturnListOfAllStimSets is not used.

Fix both issues by using ReturnListOfAllStimSets. This makes it now also
possible to use search strings together with third party stimsets.