Refactor GetTTLStimSets to be hardware independent
The current GetTTLStimSets forces the user to know a lot about the
hardware as the channel parameter is either the *hardware channel* for
ITC hardware or ignored for NI hardware.

Instead the user usually wants to just get a list of TTL stimsets for
the DAEphys TTL channels.

Refactor the function to return exactly that and also add some testing.