DC_GetChannelTypefromHS: Handle inactive headstages properly
Since 0f5cac24 (P_PressureMethodPossible: Handle TP during DAQ as well,
2019-02-18) we use DC_GetChannelTypefromHS to figure out if we can use
the pressure methods or not.

But we call this function also for the currently selected headstage for
pressure control and this might just be not active.

AFH_GetDACFromHeadstage returns NaN for inactive headstages so we need
to handle that before AFH_GetITCDataColumn correctly complains.