SplitSweepIntoComponents/SplitTTLWaveIntoComponents: Allow tuning the TTL rescaling behaviour
For ITC data the TTL waves contain up to 4 individual TTL bits.
Therefore the value ranges of these are not [0, 1] by default but
instead go up to 8.

But that is confusing for the users so it would be nice to rescale that
to the [0, 1] range. Switching unconditionally does not work as we don't
want to change the TTL data stored to NWB.

So we introduce the two flags TTL_RESCALE_ON/TTL_RESCALE_OFF which allow
different users to adapt to their needs. For the
SweepBrowser/DataBrowser we rescale, for NWB export not.