State Machine for Formula Parsing
Add formula serialization for different operations.
The formula style follows [json logic](http://jsonlogic.com/)

This is an infinite state machine using a loop to tokenize the formula
pattern.

Supported operations currently are

* addition / subtraction
* multiplication / division
* bracket evaluation (functions, parenthesis)
* functions min/max

All operations allow multi-column input waves as inputs.

The order of operations is respected and multiplication is executed after
addition. This is currently done with a look-ahed to see if a higher
operation is present in the following chars and evaluation is delayed if
so. This might not hold for all cases and is a simple work-around.

The order of calculation inside, before, and after brackets needs to be
respected. This is done with lookaheads to search for ending brackets.

A simple differentiation between functions and brackets is performed.
Functions start with alphabetical characters and end with a ")" normal
closing bracket.

The two JSON-logic functions`min`, and `max` serve as tests functions.
They accept an array of arbitrary length as function parameters.