CA_TemporaryWaveKey: Add generic wave cache accessor
A convenient way of executing code in multiple threads is

Multithread junkWave = DoSomeStuff(p, q)

where junkWave has to have a specified size. As the contents of the junk
wave don't matter, only the sizes, do, it makes sense to reuse a cached
wave for better performance. CA_TemporaryWaveKey allows to access that
cached wave.