SCOPE_ITC_AdjustFIFOPos: Adapt for decimated oscilloscope wave
The sizes of the HardwareDataWave and the Oscilloscope wave don't match
anymore with decimation.

In addition we also rework how we handle end of DAQ signalling for ITC
hardware. We used to handle a fifoPos of NaN in SCOPE_ITC_AdjustFIFOPos
by just passing it through. This works in principle now as well, but
creates problems with updating the last part of the hardware data.

Our current work around was to rewrite the Oscilloscope data wave once
we are finished with DAQ from SWS_SaveAcquiredData. But that is of
course not optimal, as it takes a lot of cycles.

We now convert a fifoPos of NaN to the last valid point in the wave so
that the scaled data/oscilloscope waves can be updated properly.

With that approach we can just use the scaled data wave for storing the
sweep data.

An important thing to notice is that SCOPE_ITC_AdjustFIFOPos is *only*
called for DAQ and not TP.