Better State/Transition/Action Separation
This approach allows a better separation of the state machines states and
actions and removes the lookahead.

The lookahead is replaced by the new `ACTION_HIGHERORDER` which inverts
the current order of calculation. This only works if the current token
evaluation is kept as linear as possible without branching in the
current recursion. This means that tokens need to be sent to recursion
before doing the inversion.