SWS_GetChannelGains: Introduce timing parameter
It is currently difficult to understand how the gains are applied for
different hardware types. Add some documentation for it and add a
timing parameter which tells us if we are concerned with the gains for
starting DAQ or after we have acquired data.

Default to timing = GAIN_BEFORE_DAQ as this does not change
anything, the parameter is also not really optional, we just use an
optional parameter for better readability.