PSQ_Ramp: Fix level calculation for NI hardware
Since the introduction of this analysis function in
3a7ef027 (PatchSeq Ramp: Add analysis function, 2018-03-20) we are using
the channel gains to convert the level in volts to the same units as
used in the HardwareDataWave.

Unforunately that approach only works if the
GAIN_BEFORE_DAQ/GAIN_AFTER_DAQ factors are the same for the hardware.

For NI hardware this is not the case, so we need to use the correct
GAIN_AFTER_DAQ here.