TP_CheckIfTestpulseIsRunning: Fix it by using the correct index
Ever since 916f26e4 (TPStorage: Keep all data in one wave, 2018-10-17)
the check that the TPStorage index is larger than the request number of
cycles is wrong as we don't create a new TPStorage wave on every TP start.

Let's introduce a new index to the TPStorage wave note which tracks the
NOTE_INDEX when starting TP. We can then use the difference between the
two to judge if we are done or not.