Allow White Spaces and Comments
Actions are preserved during the execution of the for loop until a state
transition has happened.

White spaces change the action permanently from collect to skip.
Comments stop the evaluation of the state transition until a newline
state has been encountered.

Multiline Input for notebook in BSP:

A new notebook is created for inputting the sweepformula. The notebook
allows plain text input. The content is read directly without processing
to a set variable. The setvariable is used for displaying the
parsing status in later versions of this panel.