Introduce scaledDataWave
Until now we always used the oscilloscope data wave for getting access
to the acquired and scaled data during DAQ/TP.

This will break down once we introduce decimation for the oscilloscope
wave for DAQ. And also for getting TP info, storing the TP wave we used
to just use the oscilloscope wave.

Therefore we introduce a scaled data wave.

This wave has scaled data up to the current FIFO position, and because
we create it independently of hardware constraints, we can just use it
after DAQ for the acquired data.