Add rms function
The root mean square is evaluated row based if no column is available
and column based otherwise. Due to transposing the wave it is not
possible to predict the results for higher dimensions. An ASSERT is
therefor introduced to avoid undefined behavior.
The result is then tested against the return value from wavestats.