Allow Single Signed Operations
Single Operand Operations like -1 and +1 are now supported. Note that
`+1` is not a valid JSON String. For simplicity reasons, -1 is evaluated
as an operation on 1 and not as a number. This also means that the first
element in an input array like [-1,0,1] will always get expanded by the
formula parser.