RA_Counter/RA_CounterMD: Handle stopped DAQ properly
We push the calls to the function RA_Counter* to the
operation queue via ExecuteListOfFunctions in RA_HandleITI*.

And if you now right after leaving the function RA_HandleITI* stop DAQ
we will call RA_Counter* without DAQ being running.

But we should handle that gracefully as well.

The error message the user sees in this case is

Message: "Uninitialized raCycleID detected"

from DC_PlaceDataInHardwareDataWave. And this is actually an incarnation
of another bug reported earlier (issue 141).

Closes #141.