AFM_CallAnalysisFunctions: Define lastKnownRowIndex/lastValidRowIndex more carefully
Until now we just used fifoPosition to fill in
AnalysisFunction_V3::lastKnownRowIndex.

But our internal management of fifoPosition should not be exposed to the
analysis function users.

So let's initialize it properly for the cases we know how to handle.

In addition we now also fix an off-by-one issue as we use the
lastKnownRowIndex/lastValidRowIndex as zero-based indizes and not as
length.