OverlaySweeps: Allow to only overlay the sweeps from the current RAC/SC
Add two checkboxes which allow the user to limit the overlayed sweeps to
ones from the same repeated acquisition cycle (RAC) or the same stimset
cycle.

The order of checking is RAC and then SC, as we always have a new RA
when a new SC starts.

A new SC starts if it starts for one headstage.