Rework DC_ChannelIsActive to DC_GetFilteredChannelState
DC_ChannelIsActive was introduced in 8256a9c3 (Change channel usage for
Testpulse, 2015-10-14) to allow querying the channel state respecting
TP/DAQ specialities.

While that worked pretty well, the calling sites had to handle all the
complexity of handling the additional status wave.

But if we let the function return the result for all channels, similiar
to what we do in DAG_GetChannelState, the call sites look much better
and we are also prepared for taking into account more channel
properties.