WB_MakeWaveBuilderWave: Add duration entry for combined stimset
Since e5c450b5 (Tests for Epoch information gathering on various
stimsets, 2019-06-25) we write epoch information into the labnotebook.

In DC_AddEpoch we therefore fetch the epoch duration for each epoch of
each sweep of the stimset.

But we did not have a duration for Wavebuilder stimsets of type
combined. So let's add the duration here as well.