Indexing: Use 0-based indizes
Since forever we have used 1-based indizes for the indexing storage
waves. This was a mere historical artefact and made adapting the code
quite confusing.

This commit changes that to use 0-based indizes. In addition we also use
now a single indexing storage wave for both DA/TTL channels and use
dimension labels for accessing it.