DAEphys: Fix pressure device listing
In 66fca8ab (DAEphys: Rework device selection logic, 2019-10-08) we
replaced all calls of HW_ITC_ListDevices with a call to
DAP_GetITCDeviceList. And in this case here we wanted to be extra smart
and just use DAP_GetDACDeviceList.

But for the updating the list of pressure devices we don't need the list
of devices capable of DAQ but all unlocked devices.