Update Introduction To LabNotebooks
The introduction to the structure of the *labnotebooks* is re-written to
avoid duplication. It clarifies the structure and uses `array` for the
meta data stored in the labnotebook and `storage containers` for the
waves or datasets. Such differentiation between structure and actual
storage limitation makes clear that the data storage is independent from
the concept of a labnotebook.