Allow Reading Labnotebook entry in SweepFormula.
Use the labnotebook of the currently loaded window to gather information
about specifically selected strings.

Issue #302

squash! Allow Reading Labnotebook entry in SweepFormula.

Let the readout of LabNotebook(s) in SweepFormula be aware of channels.

First, the channel information is retrieved from the labnotebook. If the
channels are not found, `labnotebook()` will bail out.  The entry for
the given key is searched in the labnotebook and evaluated for the
associated channel.

The function is aware of headstage independent values.

Usage:

`labnotebook(string key, array channels, array sweeps)`