DC_PlaceDataInHardwareDataWave: Make TP work again with NI hardware
Since 32f55750 (DC_PlaceDataInHardwareDataWave: Cache the HardwareDataWave for the Testpulse, 2019-10-25)
we cache the hardware data wave for TP.

But we failed to realize that we not only get a free wave from the wave
cache but also the referenced waves therein are free waves. And NI
hardware requires us to use permanent waves so this currently bugs out.

If we initialize the DAQDataWave with the correct permanent NI channel
waves we can use the new recursive mode of MoveWaveWithOverwrite.

Report by Michael Huth.