Document specialities about unassociated entries from more
Unassociated DA/AD labnotebook entries are suffixed with
"UNASSOC_$channelNumber". At first sight this might indicate that the
entries are not distinct if you have multiple unassociated DA/AD
channels with the same channel number.

But it turns out that we don't have labnotebook entries which have the
same name for the DA and AD channels.

In order that it also stays like that we add a small test to check that.