fix for AbortTP test failing, DQ_ApplyAutoBias referred to non existing panel
When the DA_Ephys panel is killed while TP is running or TP is stopped then
on next TP start the remaining data is read out from Async.
This means the panel can already have been closed or reopened.

The fix skips the Auto Bias apply if the window is not locked any more.

However, if in regular operation a user opens a new DA_Ephys panel and locks
it to the same device, DQ_ApplyAutoBias will work with the current GUI state
when referring to the old TP data.