Varying Channel Numbers for Sweeps in SweepFormula
Generate a *unique* list of input channel types and numbers from the
`channels` and `sweeps` parameters. The output wave contains all
channels as columns and sweep numbers as rows. The column labels specify
the channel type and number.

+--------+----------+----------+----------+
|        | channel1 | channel3 | channel9 |
+--------+----------+----------+----------+
| sweep1 |    data  |    data  |  data    |
+--------+----------+----------+----------+
| sweep2 |    data  |          |  data    |
+--------+----------+----------+----------+
| sweep3 |          |          |  data    |
+--------+----------+----------+----------+

Missing fields are filled with `NaN`. This means that if i.e. `sweep3`
is only associated with `channel9` but `channel1` and `channel3` are
requested, the missing information is filled with `NaN`. As a usual
labnotebook entry is not allowed to contain `NaN`, such entries always
indicate that information is missing.

If the user requests `channels(AD,AD1)` and `AD1` is already contained
in `AD`, the channelNumber is ignored.

It is possible to mix channelTypes: `channels(DA,AD)`. In this case, the
sort order is AD,DA and follows the constants `ITC_XOP_CHANNEL_TYPE_ADC`
and `ITC_XOP_CHANNEL_TYPE_DAC`.