CreateLBNUnassocKey: Use a new naming style for unassociated channels
The old style was ambigous as it did not encode the channel type.
Although this does not create a problem in theory, it is in practice
difficult to know if a setting is from an AD or a DA channel.

The new style solves these issues.

For the buggy-data-fixup in NWB_AppendSweepLowLevel we need to query
only old style entries.