AI_FillAndSendAmpliferSettings: Avoid storing amplifier settings for non-matched clamp mode
If the user changes the clamp mode during a sweep in the MCC
application, the clamp mode returned by AxonTelegraph_DataStruct is
different from the clamp mode MIES thinks it has.

And that leads to unexpected labnotebook entries being recorded.

Assert out for now until we figure out a better solution.