OOD_CalculateOffsets: Prevent changing the very first region
The OOdDAQ algorithm was adapted in 88323d8d (Replacement of oodDAQ
offset calculation routines, 2019-06-13). This changed the internal
algorithm completely. The function OOD_CalculateOffsets calculates the
required offsets from the existing regions. For compatibility with yoked
devices it supports a preload wave called baseRegions. For non-yoked
execution or the first call of yoked execution this baseRegions wave is
initialized with the very first region.

But that initialization must happen by duplicating the wave not reusing
the wave, as we later on resize baseRegions and add more entries to it.

We write the calculated regions to the labnotebook as "oodDAQ regions".

Buggy regions look like

  '_free_'[0]= {"500-1506.01;1506.01-2512.02;","1506.01-2512.02;","2512.02-3518.03;","3518.03-4524.04;"}

where as they now are

  '_free_'[0]= {"500-1506.01;","1506.01-2512.02;","2512.02-3518.03;","3518.03-4524.04;"}

Although we do have regression tests for oodDAQ the buggy code path was
not hit as we only tested with two stimsets.

Raising the cache key version is also required as we now have buggy
regions in the cache for oodDAQ.