SweepFormula: Update data() function
The data function is updated. It collects a unique list of channels
first and then sends it to SF_GetSweepForFormula for creating a
3dimensional wave.

The individual sweeps can have different starting values and the
complete range is determined before data is put into the wave. This
allows to have empty passages within the data() function.