Require latest IP8 nightly version
This fixes a bug which can be more-or-less easily triggered when the
databrowser/sweepbrowser is open, a formula plot is open and the databrowser/sweepbrowser
is closed.

It does not immediately crash but might also hang on for a while.

The reason is that the following code

Function crash()

	NewDataFolder/O/S foo
	make ddd
	display ddd
	SetDataFolder root:
	killwrapper("root:foo")
End

ThreadSafe Function killwrapper(String dfPath)
	KillDataFolder/Z $dfPath
End

kills the ddd wave for old IP versions. But that wave is displayed and
thus must not be killed.