Analysis function params: Allow arbitrary parameter values
Since the initial introduction of analysis parameter support in 55ee1ccf
(Introduce Analysis functions V3, 2018-01-26) we did not allow certain
characters like ";:=,|" to be present in the string or text wave
contents.

The reason for that restriction was that we needed to be able to parse
the contents using Igor style key value pairs from the textual
wavebuilder wave and the stimset wave note.

But that restriction was quite limiting.

Therefore we now use percent encoding for the parameter values, which
prevents enforcing limitations on the payload.