Add Unit Test For SweepFormula
The Unit Test builds up a fake DataBrowser graph with traces and a
corresponding notebook entry. It tests first if all channels can be
retrieved from the notebook as this function is used by the labnotebook()
and data() function. After this test, a custom data entry is retrieved
in `UNKNOWN_MODE` and the input data set is returned from the graph and
checked for equality.