CONF_RestoreHeadstageAssociation: Avoid passing NaN amplifier serials to AI_OpenMCCs
Amplifiers are currently disabled by setting them to null in the JSON
config file. This null is mapped to NaN [1] in the JSON XOP.

Currently we also gather these NaNs in ampSerialList and pass that to
AI_OpenMCCs. In case the MCC application is not yet open, this creates
bogus MCC apps as the serial NaN is not known.

Let's just skip over the NaNs when gathering the list entries.

[1]: https://docs.byte-physics.de/json-xop/api.html#booleans