GetInteractiveMode: Use a thread dependent default
We recently had some crash reports where the interactiveMode global
variable showed up. While nothing is per se wrong how we are doing things,
we can be a bit more careful here. The reason is that we want interactive
mode only to be true in the main thread and not in preemptive threads. For
example HW_ITC_GetZValue() uses the interactive mode to enable/disable
error pop ups, and we definitly don't want error popups in preemptive
threads.