BSP_DynamicStartupSettings: Use correct order of tab calls
In f06aa8d0 (Databrowser/SweepBrowser: Use better defaults for the
SweepFormula tab, 2019-11-27) we ensured that the sweepformula formula
tab is the selected one after locking.

But we need to do that before setting the main tab to Settings as
otherwise the formula notebook is shown on the Settings tab.