Fix the TP power spectrum
This is broken since a unknown time.

The reason is that FFT does not allow to combine /COLS with the /OUT
flag. IP8 silently sets the output to NaN, only IP9 complains at
runtime.

We fix that by using the TPOscilloscopeData wave for displaying the
power spectrum and manually converting the FFT result to real.