Labnotebook: Find DAQ entries better with sweep rollback and ancient labnotebooks
Consider a labnotebook which does not have an entry source type. And
the user then uses sweep rollback, runs TP and reacquires the sweep.

This results in a labnotebook like

Row | Sweep | (Actual entry source type) | SomeEntry
----|-------|----------------------------|----------
1   | 73    | DAQ                        | 1
2   | 74    | DAQ                        | 2
3   | 74    | DAQ                        | 3
4   | 73    | TP                         | NaN
5   | 74    | DAQ                        | 4
6   | 74    | DAQ                        | 5
----|-------|----------------------------|----------

When we now try to query the labnotebook info for 73, "SomeEntry" and
DATA_ACQUISITION_MODE we don't find anything as we think that row 4 is
the correct match. The reason is that we don't have entry source type
information so we can't just ignore row 4.

The bugfix is now to use a heuristic to determine the entry source type
of Row 4. We only do that for the common case that DATA_ACQUISITION_MODE
is requested. As in this case we can remove all rows which have
entries for "TP Peak Resistance" or "TP Baseline Fraction", as these are
definitly TP only.

We have the entry source type since 1b21f99c (Bugfix: Disambiguate
labnotebook entry search, 2016-10-08), but can't use the heuristic for
old labnotebooks implemented therein as with sweep rollback we have
distinct ranges we have to examine earlier on.