ITC Hardware: Convert XOP error codes to the expected constants
The constants defined in an XOP for the error codes are transformed so
that they don't overlap between XOPs.

We therefore need to transform them back.

This was forgotten in 8d3aebd5 (Switch to ITC XOP2, 2016-05-31). But
never really surfaced as XOP errors from the ITC XOP2 are very rare.