Fix cache stats wave upgrade path
In 8bd2f453 (Cache: Create the statistics wave with the correct size,
2018-10-05) we fixed the upgrade path for the cache stats wave.  This
fix worked for the case of differing number of rows of the keys/values
wave and the stats wave. But fails for the case where the sizes of all
three waves are the same.

We now do the right thing in this case as well. In addition we also
clear the new space in the wave after resizing.

Another fix is that we now also write the correct NOTE_INDEX into the
wave note.