Databrowser: Fix transition from no sweeps to first acquired sweep
With the new configuration we can have the situation that the
databrowser should update on new data for the very first sweep already.

But that did not work as the check formerLast != last in
DB_UpdateLastSweepControls is never true for the first time as both are
zero.

We can fix that by having the default value for valdisp_SweepControl_LastSweep be NaN.