ExperimentDocumentation: Speedup querying labnotebook entries for the changed wave note entries
Since 7aee1b63 (Add a cache for the labnotebook queries, 2018-05-28) we
cache the rows of a sweep in the labnotebook. But that is, correctly,
only done if we have found an entry in the labnotebook. The way the
changed entry detection in
ED_WriteChangedValuesToNote/ED_WriteChangedValuesToNoteText works we
usually have a lot of non-existing entries before we have the first
existing ones.

This results in around 140 calls to FindRange which is way too much.
By searching for an entry which always exists first, we can fill the row
cache and avoid unnecessary calls to FindRange.