GetListOfObjects: Rewrite and speed it up a lot
The current implementation of GetListOfObjects suffers from N^2
complexity as it uses GetIndexedObjNameDFR.

This can be worked around by using VariableList/StringList/WaveList and DataFolderDir.

We also drop expensive flags like matchList and waveProperty. matchList
can be replaced with ConvertListToRegexpWithAlternations, waveProperty
does not have a replacement.

One drawback of the new approach is that GetAllObjects now fails if the
datafolder has a comma `,` in it. But this limitation is documented and
not a problem for us.